Re-enabled stabs reading in the kernel debugger.
In ke/main.c, stabs for boot time drivers that were loaded with the kernel
are linked and enabled before calling KdbInit, so that the user has them
available.

In ldr, drivers are now loaded with their symbols if available.  This applies
to userland dlls, too.

I added DEL as well as backspace to the keys that cause the cursor to back
up in kdb.

The 'addr' command in kdb now prints a correct address (with symbol
info if available).

The kernel stack trace printer in exp.c now uses kdb to print addresses if
it was enabled.


git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@7598 97493ccd-5924-5043-b1f5-66cb403b36ce