Brazilian Portuguese translation for regedit by Marcelo Zamperetti (marcelohz AT gmail DOT com) For the record, this seems to be the first program that we now have in both Brazilian Portuguese and Portuguese from Portugal :-)
If someone has time, please look at the regedit code and find out which dialogs are really still needed.
For me it looks like the layout of the resource files was changed some time ago and now we have files with the old and with the new layout.
See issue #2440 for more details.


git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@27832 97493ccd-5924-5043-b1f5-66cb403b36ce