[NTOSKRNL] Previously, in ReactOS' stories: ReactOS was broken since more a hundred of commits when two devs decided it was enough. One (we will name him Timo) fixed the context switch. The second, with the help of a third dev (Pierre & Hervé) decided to run tests by disabling some stuff (ie INIT_FUNCTION). This magically made testbot going back to life. But, some complains came from the ML, so trying to appease world, Pierre decided to revert half of his changes, thinking this would be OK.
~~
Now, after a night. It appears that half revert was a pretty bad idea. Testbot keeps being broken.
So, this commit reverts r49665 (which was half reverting r49662). That way, testbot should be back, able to run tests. BUT, due to a NPFS issue, some tests are broken. Eric has been nicely mailed about that issue, with an idea of fix (thanks go here to Aleksey).

~~
For those who like that, drama to follow on ML.
I turn into being S/M...

git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@49691 97493ccd-5924-5043-b1f5-66cb403b36ce