[ADVAPI32] Enable the new SAM-based logon system.
Using your own accounts to log-on to ReactOS will still fail because of bugs in winlogon, userenv.dll and the registry. But at least the infrastructure is in place and we can start to fix the reamining bugs. All accounts still use a hard-coded privilege set, but this will be fixed as soon as possible.

git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@60800 97493ccd-5924-5043-b1f5-66cb403b36ce