[FASTFAT] Implement part 2/3 (first part was in ntoskrnl, remember ;-)) of renaming support in ReactOS.
Here, we implement support for SL_OPEN_TARGET_DIRECTORY for IRP_MJ_CREATE in fastfat. The code is not the best ever, but we can't do miracles with fastfat.
This fixes a few tests in KmTests:IoCreateFile and also shows failures I still don't get and that will require deeper attention.

Last part will be to effectively bring support of FileRenameInformation in IRP_MJ_SET_INFORMATION. But this is yet another story.


git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@63034 97493ccd-5924-5043-b1f5-66cb403b36ce