[NTFS] Quickly implement NtfsReadFile().
By quickly, I mean that it works but is under optimal in many ways, and also doesn't support any caching. So, don't expect breaking performances.

BUUUUUUT... This implements reading a file on a NTFS volume on ReactOS! And it works!
Here is a picture of all the achievements of ReactOS with NTFS lately: http://www.heisspiter.net/~Pierre/rostests/NTFS_all.png
-> Volume information
-> Displaying files in explorer
-> Moving across directories
-> Reading a file (with more here)

That's all folks! (for now ;-))


git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@65192 97493ccd-5924-5043-b1f5-66cb403b36ce