[NTFS] - Implement NtfsGrabFCBFromTableById() and NtfsGetFCBForFileById(). These two are responsible for opening and retrieving a file open by ID. - Modified NtfsMakeFCBFromDirEntry() so that it supports openings by ID which come with less information than normal openings. - Modified NtfsGrabFCBFromTable() so that it doesn't return a FCB which would have been opened by ID. - Modified NtfsOpenFile() so that it allows opening a file by ID using the two FCB helpers functions defined upper. - Modified NtfsCreateFile() so that it can validate input in case of opening by ID.
To make it short, all this implements file opening by ID in NTFS driver. I made the choice to separate FCB which are coming from "normal" opening and these from ID opening.
The reason is that a file opening with ID has less information than a normal opening, and may fail on these information retreival. This is due to the fact that if opening by ID can virtual open all the files, because it ignores authorizations, traverses checks, in the end for retrieving the information, such checks can be done and fail.

I also left all this disabled so far, because it appears to regress NFI. I will investigate this.

CORE-8725


git-svn-id: svn://svn.reactos.org/reactos/trunk/reactos@65652 97493ccd-5924-5043-b1f5-66cb403b36ce